For conda builds, drop py36, add py39. Add numpy 1.19, 1.20.
riptable effectively dropped support for py36 in 49037d2, so the removal of
py36 support here is for consistency. py36 is also nearing EOL and isn't
supported by numpy 1.20 onward.

Begin targeting numpy 1.19 and 1.20 as well to provide users some
flexibility in which versions of numpy they want to install, and also for
compatibility with pnumpy.